tfs.lua split -> tfs.lua + tfs_api.lua
